merge small into main

+++ app.py
@@ -1,31 +1,28 @@
+import os
 import io
 import json
-import os
 import uuid
 
 from flask import Flask, request, Response
 from tempfile import NamedTemporaryFile
 from faster_whisper import WhisperModel
 
-app = Flask(__name__)
-app.config['DEBUG'] = True
+server = Flask(__name__)
+server.config['DEBUG'] = os.environ['DEBUG']
 
-MODEL_SIZE = 'small'
-UPLOAD_FOLDER = 'uploaded_files'
-ALLOWED_EXTENSIONS = {'wav','mp3'}
+modelPath = os.environ['MODEL_SIZE']
 
-@app.route('/', methods=['GET'])
+@server.route('/', methods=['GET'])
 def home():
     return {"success": True, "message": "Hello World from VTT FW"}
 
 def allowed_file(filename):
     return '.' in filename and \
-           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+           filename.rsplit('.', 1)[1].lower() in os.environ['ALLOWED_EXTENSIONS']
 
-@app.route('/test_mp3', methods=['GET'])
+@server.route('/test_mp3', methods=['GET'])
 def test_mp3():
-    model_size = "small"
-    model = WhisperModel(model_size, compute_type="int8")
+    model = WhisperModel(modelPath, compute_type="int8")
     segments, info = model.transcribe("audiobook.mp3")
     data = ''
     for segment in segments:
@@ -33,10 +30,9 @@ def test_mp3():
     del model
     return {'success': True, 'data':data}
 
-@app.route('/test_wav', methods=['GET'])
+@server.route('/test_wav', methods=['GET'])
 def test_wav():
-    model_size = "small"
-    model = WhisperModel(model_size, compute_type="int8")
+    model = WhisperModel(modelPath, compute_type="int8")
     segments, info = model.transcribe("indonesian.wav")
     data = ''
     for segment in segments:
@@ -44,7 +40,7 @@ def test_wav():
     del model
     return {'success': True, 'data':data}
 
-@app.route('/transcript', methods=['POST'])
+@server.route('/transcript', methods=['POST'])
 def transcript():
 
     # 01 field, files Validation
@@ -65,9 +61,12 @@ def transcript():
             'message': f'Failed transcript audio to text, File Name {audio_file.filename} Not Allowed',
         }
 
+    if not os.path.exists(os.environ['UPLOAD_FOLDER']):
+        os.makedirs(os.environ['UPLOAD_FOLDER'])
+
     # 03 Write unique file
     tempName = str(uuid.uuid4())
-    tempFile = os.path.join(app.config['UPLOAD_FOLDER'], f'{tempName}_{audio_file.filename}')
+    tempFile = os.path.join(os.environ['UPLOAD_FOLDER'], f'{tempName}_{audio_file.filename}')
     
     # 04 Read Open & Write audio data to file uploaded files
     audio_data = audio_file.read()
@@ -75,7 +74,7 @@ def transcript():
         f.write(audio_data)
 
     # 05 load model and start transcribe audio to text
-    model = WhisperModel(MODEL_SIZE, compute_type="int8")
+    model = WhisperModel(modelPath, compute_type="int8")
     segments, info = model.transcribe(tempFile)
     data = ''
     for segment in segments:
@@ -89,16 +88,14 @@ def transcript():
 
     return {'success': True, 'data':data}
 
-@app.route('/clear_uploaded_files', methods=['GET'])
+@server.route('/clear_uploaded_files', methods=['GET'])
 def clear_uploaded_files():
-    paths = app.config['UPLOAD_FOLDER']
+    paths = os.environ['UPLOAD_FOLDER']
     lenFile = 0
-    for f in os.listdir(paths):
-        os.remove(os.path.join(paths, f))
-        lenFile += 1
+    if os.path.exists(paths):
+        for f in os.listdir(paths):
+            os.remove(os.path.join(paths, f))
+            lenFile += 1
 
     removedStr = str(lenFile)
     return {'message' : 'Removed ' + removedStr + ' Files'}
-
-if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=5000)
